show_filesize/requirements.txt
GitPython==2.1.10
natsort==5.2.0
